+++ kernels/moe/mixed_moe_gemm_2stage/__init__.py
@@ -1,0 +1,15 @@
+# SPDX-License-Identifier: Apache-2.0
+# Copyright (c) 2025 FlyDSL Project Contributors
+
+"""Mixed-dtype MoE 2-stage MFMA kernels (stage1 / stage2).
+
+Split from the former monolith; public API unchanged.
+"""
+
+from kernels.moe.mixed_moe_gemm_2stage._gemm1 import compile_mixed_moe_gemm1
+from kernels.moe.mixed_moe_gemm_2stage._gemm2 import compile_mixed_moe_gemm2
+
+__all__ = [
+    "compile_mixed_moe_gemm1",
+    "compile_mixed_moe_gemm2",
+]
