+++ kernels/moe/mixed_moe_gemm_2stage/_common.py
@@ -1,0 +1,46 @@
+# SPDX-License-Identifier: Apache-2.0
+# Copyright (c) 2025 FlyDSL Project Contributors
+
+"""Mixed-dtype MoE 2-stage MFMA kernels (stage1 / stage2). :: _common"""
+
+import os
+
+from flydsl._mlir.dialects import llvm
+
+
+def _get_cu_num() -> int:
+    env = os.environ.get("CU_NUM")
+    if env:
+        return int(env)
+    try:
+        import torch
+
+        return int(torch.cuda.get_device_properties(torch.cuda.current_device()).multi_processor_count)
+    except Exception:
+        return 304
+
+
+def _barrier(vmcnt=63, lgkmcnt=63):
+    """Emit s_waitcnt + s_barrier via inline asm.
+
+    Bypasses LLVM SIInsertWaitcnts which would insert a conservative
+    s_waitcnt vmcnt(0) lgkmcnt(0) before every S_BARRIER MI.
+    """
+    parts = []
+    needs_waitcnt = vmcnt < 63 or lgkmcnt < 63
+    if needs_waitcnt:
+        wc = []
+        if vmcnt < 63:
+            wc.append(f"vmcnt({vmcnt})")
+        if lgkmcnt < 63:
+            wc.append(f"lgkmcnt({lgkmcnt})")
+        parts.append("s_waitcnt " + " ".join(wc))
+    parts.append("s_barrier")
+    llvm.InlineAsmOp(
+        res=None,
+        operands_=[],
+        asm_string="\n".join(parts),
+        constraints="",
+        has_side_effects=True,
+        is_align_stack=False,
+    )
