+++ kernels/moe/moe_blockscale_2stage/__init__.py
@@ -1,0 +1,25 @@
+# SPDX-License-Identifier: Apache-2.0
+# Copyright (c) 2025 FlyDSL Project Contributors
+
+"""MoE block-scale 2-stage MFMA kernels (stage1 / stage2 / reduction).
+
+Split from the former monolith; public API unchanged.
+"""
+
+from kernels.moe.moe_blockscale_2stage._gemm1 import compile_moe_blockscale_gemm1
+from kernels.moe.moe_blockscale_2stage._gemm2 import (
+    MoeGemm2Mode,
+    _MoeGemm2ReduceWrapper,
+    compile_moe_blockscale_gemm2,
+    compile_moe_blockscale_gemm2_ex,
+)
+from kernels.moe.moe_blockscale_2stage._reduction import compile_moe_reduction
+
+__all__ = [
+    "compile_moe_blockscale_gemm1",
+    "compile_moe_blockscale_gemm2",
+    "compile_moe_blockscale_gemm2_ex",
+    "compile_moe_reduction",
+    "MoeGemm2Mode",
+    "_MoeGemm2ReduceWrapper",
+]
